move observer to discourseGraphsMode
